♻️ refactor backend main routes
