+++ backend/routes/misc_routes.py
@@ -1,0 +1,7 @@
+from fastapi import APIRouter
+
+misc_router = APIRouter()
+
+@misc_router.get("/")
+async def root():
+    return {"status": "OK"}
